Move a few functions implementations from quaternion.h to quaternion.cc
